ia_training_rag_custom/requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
blis==1.3.3
brotli==1.2.0
build==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.3.1
cloudpathlib==0.23.0
colorlog==6.10.1
confection==1.3.3
cryptography==46.0.5
cssselect2==0.8.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.13
data==0.4
dataclasses-json==0.6.7
datasets==4.8.2
decorator==5.2.1
Deprecated==1.3.1
dill==0.4.1
distro==1.9.0
docling==2.73.0
docling-core==2.64.0
docling-ibm-models==3.11.0
docling-parse==4.7.3
durationpy==0.10
et_xmlfile==2.0.0
faiss-cpu==1.7.4
Faker==40.4.0
fastapi==0.128.8
filelock==3.20.3
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.61.1
fr_core_news_md @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.8.0/fr_core_news_md-3.8.0-py3-none-any.whl#sha256=8a70d090a54ef77525c3ffa6a6195b9d365f2cf369ae1cd84ede93f3d709079e
fr_dep_news_trf @ https://github.com/explosion/spacy-models/releases/download/fr_dep_news_trf-3.8.0/fr_dep_news_trf-3.8.0-py3-none-any.whl#sha256=df243e539bbc365390744a411aa574fb03eb2ee2f948cf6149ad722c59ab419b
frozenlist==1.8.0
fsspec==2026.2.0
funcsigs==1.0.2
future==1.0.0
gitdb==4.0.12
GitPython==3.1.46
googleapis-common-protos==1.72.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
importlib-metadata==6.11.0
importlib_resources==6.5.2
iniconfig==2.3.0
Jinja2==3.1.2
jiter==0.13.0
joblib==1.5.3
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langchain==0.3.0
langchain-community==0.3.0
langchain-core==1.2.12
langchain-ollama==1.0.1
langchain-openai==1.1.9
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.7.1
latex==0.7.0
latex2mathml==3.78.1
lxml==4.9.3
markdown-it-py==4.0.0
marko==2.2.2
MarkupSafe==2.1.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
murmurhash==1.0.15
mypy_extensions==1.1.0
narwhals==2.16.0
networkx==3.6.1
numpy==1.26.4
oauthlib==3.3.1
ocrmac==1.0.1
ollama==0.6.1
omegaconf==2.3.0
onnxruntime==1.24.1
openai==2.20.0
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.11.7
overrides==7.7.0
packaging==23.2
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.4
pillow==10.4.0
pluggy==1.6.0
polyfactory==3.2.0
posthog==7.8.6
preshed==3.0.13
propcache==0.4.1
protobuf==3.20.3
psutil==7.2.2
pulsar-client==3.10.0
pyarrow==23.0.0
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pydyf==0.12.1
Pygments==2.19.2
pylatexenc==2.10
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-CoreML==12.1
pyobjc-framework-Quartz==12.1
pyobjc-framework-Vision==12.1
PyPDF2==3.0.1
pypdfium2==5.4.0
pyphen==0.17.2
PyPika==0.51.1
pyproject_hooks==1.2.0
pytesseract==0.3.10
pytest==9.0.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
rapidocr==3.6.0
referencing==0.37.0
regex==2026.1.15
reportlab==4.0.7
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.30.0
rtree==1.4.1
ruamel.yaml==0.19.1
ruff==0.15.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
semchunk==2.2.2
sentence-transformers==5.3.0
sentencepiece==0.2.1
shapely==2.1.2
shellingham==1.5.4
shutilwhich==1.1.0
six==1.17.0
smart_open==7.5.1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.14
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.46
srsly==2.5.3
starlette==0.52.1
streamlit==1.55.0
sympy==1.14.0
tabulate==0.9.0
tempdir==0.7.1
tenacity==8.5.0
thinc==8.3.13
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.3
transformers==4.57.6
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
typer==0.21.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
uvloop==0.22.1
validators==0.35.0
wasabi==1.1.3
watchfiles==1.1.1
weasel==1.0.0
weasyprint==68.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
wrapt==1.17.3
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zopfli==0.4.0
zstandard==0.25.0
